dwtest: add explicit testpoint for runtime.throw
Add a new testpoint that verifies the DWARF var location for the
incoming parameter of the function 'runtime.throw'. We already have
testpoings that validate DWARF for regular user-written functions that
take a single string argument (as does 'runtime.throw'), but it is
good to test the function in the runtime as well, since the rules for
compiling the runtime are special, and we occasionally problems with
DWARF generation can crop up that are runtime-specific.

Fixes golang/go#62523.

Cq-Include-Trybots: luci.golang.try:x_debug-gotip-linux-amd64-longtest
Change-Id: I6558d9f12b22c997535cfc23d4b0758b10bd9548
Reviewed-on: https://go-review.googlesource.com/c/debug/+/526836
Reviewed-by: Alessandro Arzilli <alessandro.arzilli@gmail.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>