cmd/viewcore: migrate from chzyer/readline to ergochat/readline
Module ergochat/readline is a fork that is actively maintained and
includes numerous bug fixes that are not present in the original module.
Additionally, this change reduces the number of indirect dependencies by
removing the need for github.com/chzyer/logex and
github.com/chzyer/test.

Details:
- Updated imports and removed old module dependencies.
- Replaced Terminal field references with Stdout method calls for shell
output.
- Modified the cmdToCompleter function to adapt to the new
PrefixCompleter structure.
- Ensured the completion tree is built using the new SetChildren and
Children methods.

Signed-off-by: Mikel Olasagasti Uranga <mikel@olasagasti.info>