gocore: add in mpratt's feedback and remove delve dependency
This change integrates some feedback from mpratt in CL 635227 that came
in after it landed. It also removes the dependency on delve, since this
is definitely reaching into internal packages that are not intended to
be used by outsiders. Instead, we just copy over what we need. This is
safe to copy over because it's largely about immutable formats and
functionality: hardware register names and DWARF 2.

Change-Id: I5aa8cb9a56b464c0cdb45baf4ea8423989d0afbe
Reviewed-on: https://go-review.googlesource.com/c/debug/+/635281
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Pratt <mpratt@google.com>
Auto-Submit: Michael Knyszek <mknyszek@google.com>