gocore: skip nil largeType when discovering pointers
If a core is taken while a goroutine is allocating a large object (which
is more likely than smaller objects, because zeroing is delayed and
takes longer than for smaller objects) gocore might observe a nil
largeType in the span, which is left nil to prevent the GC from trying
to scan the partially-initialized object.

Fixes golang/go#71182.

Change-Id: Iafa7aed29466dc8ad6eac8f10171885421e10f76
Reviewed-on: https://go-review.googlesource.com/c/debug/+/641515
Reviewed-by: Nicolas Hillegeer <aktau@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>