gocore: fix bad processing of large object pointers
Currently large object pointer finding has a bug in that we don't
interpret largeType as an array element type (which it is) and instead
as the single type for the entire array. This means we miss lots of
pointers in large objects that are, for example, a slice of pointers.

Change-Id: I9ade197b5db9ac52b829524e0d70822c8fdc1dbf
Reviewed-on: https://go-review.googlesource.com/c/debug/+/659335
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Nicolas Hillegeer <aktau@google.com>