internal/gocore: fix pointer locations for small objects in green tea
The tests are currently broken when GOEXPERIMENT=greenteagc because that
shifted where the pointer/scalar bits are located for small objects a
little bit. This change modifies gocore to look in the right place and
detect if GOEXPERIMENT=greenteagc is enabled (and if we're using Green
Tea in general, if/when it becomes the default).

This was not caught by the builders because we don't have an x/debug
builder that sets the GOEXPERIMENT.

Change-Id: I746ce601ebc12738920616889200f23642f7280f
Reviewed-on: https://go-review.googlesource.com/c/debug/+/699975
Reviewed-by: Michael Pratt <mpratt@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Nicolas Hillegeer <aktau@google.com>
Commit-Queue: Nicolas Hillegeer <aktau@google.com>